requirements.txt
pytest>=6.2.0
pytest-bdd>=5.3.0
python-dotenv>=0.19.0